Merge `agent_id` into registration info returned by `register_agent()` (#126)
* Merge `agent_id` into `AgentRegistration`

This is a followup to PR #124 that addresses the complications
of typing the nested generic types.

The goal is to turn:

    def register_agent(
        behavior: BehaviorT, ...
    ) -> tuple[AgentId[BehaviorT], AgentRegistration]: ...

into:

    def register_agent(
        behavior: BehaviorT, ...
    ) -> AgentRegistration[BehaviorT]: ...

where `agent_id: AgentId[BehaviorT]` is an attribute of
`AgentRegistration` while preserving the fact that each
`ExchangeTransport`, and therefore factory, has its own
`AgentRegistration` type.

I did this in a few parts: First I made `ExchangeFactory` and
`ExchangeClient` generic on `ExchangeTransport`. Then I made
`ExchangeTransport` a protocol generic on `AgentRegistration`. This sort
of separates the transport+registration relationship from the
factory+transport and client+transport relationships.

With the change to a protocol and some required reordering of class
definitions, it made sense to move `ExchangeTransport` and affiliated
classes to its own module.

* Reuse TypeVar declarations