Rewrite Academy in asyncio (#128)
* Make exchanges async

* Rename thread exchange to local

* Make agent runner async

* Make agents and handles async

* Fix mypy plugin and errors

* Tests and coverage passing

* Update examples to new syntax

* Update docs and restore Manager name

* Change Handle IDs

Changes the mailbox_is attribute to client_id which is more clear
than inferring the client and attached to a mailbox. Also removes
the mailbox_id parameter to BoundRemoteExchange which could just
be inferred from the transport.

* Simplify BoundRemoteHandle to RemoteHandle

* Remove _agent_id passthrough hack

* Fix Python 3.9 compatibility

Python 3.9 doesn't handle asyncio.Event/Future/Queues created outside
of the event loop, causing errors when initializing and agent before
submitting it to be run in a different loop.

* Address bad coverage checking in Python 3.11