Defer agent initialization to workers in `Launcher` (#130)
* Defer Agent creation to on worker in Launcher

* Make Agent not pickleable

* Remove deferred Event init for Python 3.9 backport

* Move handle binding logic to academy.agent

The Agent is the only thing that uses the behavior handle binding logic
so I think it makes more sense to keep it with the Agent so the callback
complexity can be removed. This also makes it a top-level function
for easier testing.

I also found a mistake in the def of UnboundRemoteHandle.close(),
discrepancy in the read-only Handle propery attributes, and
simplified RemoteHandle.bind_to_exchange().

* Rename bind_to_client and remove from RemoteHandle

Renames `bind_to_exchange` to `bind_to_client` of
`UnboundRemoteExchange` which is more accurate to what the method does:
creates a handle bound to a specific exchange client.

Removes `bind_to_client` from `RemoteHandle`, as it was not used
anywhere in the code (except for a contrived unit test), and basically
just made a replica of the current handle.

* Simplify Agent.run()

* Fix launcher restart test